Iterators-and-Streams/siddhanth-wf-dynamic/pyproject.toml
[tool.poetry]
name = "siddhanth-wf-dynamic"
version = "0.1.0"
description = "Dynamic text processing using YAML pipeline"
authors = ["siddhanthagarwal984 <siddhanthagarwal984@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = {extras = ["all"], version = "^0.6.1"}
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dynamic = "siddhanth_wf_dynamic.commands:app"